requirements.txt
smolagents
python-dotenv
mcp
litellm
